each function has a single responsbility
